+++ tsconfig.json
@@ -19,5 +19,5 @@
     }
   },
   "exclude": ["node_modules"],
-  "include": ["./src/**/*"]
+  "include": ["./src/**/*", "./__helpers__/*"]
 }
